+++ NEWS.md
@@ -1,7 +1,6 @@
 # Newspaper for the mixcurelps package #
 
 ### Version 0.1.1 ###
-
 * **2021-09-09:** Release of unstable version on Github. Version name: "Keeping CPU at low temperature".
 
 ### Version 1.1.1 ###
@@ -11,3 +10,9 @@
 ### Version 1.1.2 ###
 * **2022-02-04:** Minor updates, adding computation of CI width in simulation
  routines. Version name: "MCMC kicks in again".
+ 
+ ### Version 1.1.3 ###
+ * **2024-10-30:** Minor update to lpsmc routine. Added two additional inputs:
+(1) initializing value for MAP search of penalty parameter and (2) binary 
+indicator to select if yes or no positive definiteness check should be done.
+Version name: "lpsmc inputs".